artmp: host multipath fixes for multiplatform nest
Preserve existing local multipath/host edits.

+++ zip_cd_entry_map.h
@@ -90,7 +90,7 @@ struct ZipStringOffset20 {
   static constexpr size_t offset_max = (1u << 20) - 1;
   static constexpr size_t length_max = (1u << 12) - 1;
   uint32_t name_offset : 20;
-  uint16_t name_length : 12;
+  uint32_t name_length : 12;
 };
 
 static_assert(sizeof(struct ZipStringOffset20) == 4);
